Update specs.Rmd
Merged Modules 1&2, so module 3 is now module 2. Replaced R code with bash code to identify the top sequence at the end of the first module (also removed the hard coding of the top sequence, now the best sequence is extracted automatically with cut and sed). module_3.R renamed to call rlp.R (for read length product)