Update specs.Rmd
Module 1 generates a complete database and a representative database. We query against the representatives in this step.
Added the 2nd module back, with commands that subset the full database to generate a genus-specific database that we query against. It now calls a new R script called generate_genus_accessions.R to get only the relevant accessions.